+++ src/app/CMakeLists.txt
@@ -75,9 +75,9 @@ set(headers
     dock_widgets/output/gt_outputdock.h
     dock_widgets/output/gt_styledlogmodel.h
     dock_widgets/output/gt_filteredlogmodel.h
-    dock_widgets/output/logfilterproxymodel.h
-    dock_widgets/output/filterheaderview.h
-    dock_widgets/output/filterpopupwidget.h
+    dock_widgets/output/gt_logfilterproxymodel.h
+    dock_widgets/output/gt_filterheaderview.h
+    dock_widgets/output/gt_filterpopupwidget.h
     dock_widgets/output/gt_outputtester.h
     dock_widgets/output/gt_taskhistorymodel.h
     dock_widgets/explorer/gt_explorerdock.h
@@ -161,9 +161,9 @@ set(sources
     dock_widgets/output/gt_outputdock.cpp
     dock_widgets/output/gt_styledlogmodel.cpp
     dock_widgets/output/gt_filteredlogmodel.cpp
-    dock_widgets/output/logfilterproxymodel.cpp
-    dock_widgets/output/filterheaderview.cpp
-    dock_widgets/output/filterpopupwidget.cpp
+    dock_widgets/output/gt_logfilterproxymodel.cpp
+    dock_widgets/output/gt_filterheaderview.cpp
+    dock_widgets/output/gt_filterpopupwidget.cpp
     dock_widgets/output/gt_outputtester.cpp
     dock_widgets/output/gt_taskhistorymodel.cpp
     dock_widgets/explorer/gt_explorerdock.cpp
