+++ src/app/dock_widgets/output/gt_filterheaderview.h
