+++ src/app/dock_widgets/output/gt_filterpopupwidget.h
