+++ src/app/dock_widgets/output/gt_logfilterproxymodel.h
