+++ src/app/dock_widgets/output/gt_outputdock.cpp
@@ -13,10 +13,9 @@
 #include "gt_logmodel.h"
 #include "gt_styledlogmodel.h"
 #include "gt_filteredlogmodel.h"
-#include "logfilterproxymodel.h"
 #include "gt_tableview.h"
-#include "filterheaderview.h"
-#include "logfilterproxymodel.h"
+#include "gt_filterheaderview.h"
+#include "gt_logfilterproxymodel.h"
 #include "gt_application.h"
 #include "gt_logging.h"
 #include "gt_outputtester.h"
